Rename PEGS to PGSE across scripts and update references in README
